apps/api/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
python-dotenv
supabase
google-generativeai
